Check for WGL_ARB_render_texture
Apparently this extension is not supported on some Intel hardware,
causing the entire initialization to fail. Workaround is to check
for the extension and avoid calling wglGetPixelFormatAttribivARB
for WGL_BIND_TO_TEXTURE_* when the extension is not supported.
A similar check should probably be done for other extensions as
well, but I currently only have time for this one.